### 2025-08-15 #8823 @editorialbot
[PRE REVIEW]: trud: An R interface to the NHS England Technology Reference data Update Distribution (TRUD) API
**Submitting author:** @rmgpanw (<a href="http://orcid.org/0000-0002-0800-2890">Alasdair Warwick</a>)
**Repository:** https://github.com/ropensci/trud/
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
