### 2025-08-15 cae2be5
add JOSS badge to readme
