+++ NEWS.md
@@ -1,3 +1,5 @@
+# trud (development version)
+
 # trud 0.2.0
 
 ## Major changes
