+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: trud
 Title: Query the 'NHS TRUD API'
-Version: 0.2.0
+Version: 0.2.0.9000
 Authors@R: c(
     person(given = "Alasdair",
            family = "Warwick",
